refactor: move mock data folder inside common module
