+++ apps/api/src/user/user.controller.spec.ts
@@ -3,7 +3,7 @@ import { UserController } from './user.controller'
 import { UserService } from './user.service'
 import { fakeRepository } from './fake.repository'
 import { User } from '@prisma/client'
-import { users } from '../mock-data/users'
+import { users } from '../common/mock-data/users'
 
 describe('UserController', () => {
   let controller: UserController
